+++ src/app/event-schedule/populate/page.tsx
@@ -1,9 +1,11 @@
-import { PopulateEventScheduleForm } from "./populate-event-schedule";
+import { NavigateToSubmitScheduledEvents } from "@/app/event-schedule/populate/navigate-to-submit-scheduled-events";
+import { PopulateEventScheduleForm } from "@/app/event-schedule/populate/populate-event-schedule";
 import { EditScheduledEventList } from "@/app/scheduled-event/edit-scheduled-event-list";
 
 function PopulateEventSchedule() {
   return (
     <main>
+      <NavigateToSubmitScheduledEvents />
       <br />
       <PopulateEventScheduleForm />
       <br />