+++ src/app/store-provider.tsx
@@ -20,8 +20,7 @@ export const StoreProvider = ({ children }: Props) => {
 
   useEffect(() => {
     if (storeRef.current != null) {
-      // configure listeners using the provided defaults
-      // optional, but required for `refetchOnFocus`/`refetchOnReconnect` behaviors
+      // Configure listeners using the provided defaults
       const unsubscribe = setupListeners(storeRef.current.dispatch);
       return unsubscribe;
     }